+++ packages/agent/src/agent/http/index.ts
@@ -150,6 +150,11 @@ export interface HttpAgentOptions {
    * Alternate root key to use for verifying certificates. If not provided, the default IC root key will be used.
    */
   rootKey?: ArrayBuffer;
+
+  /**
+   * Whether or not the root key should be automatically fetched during construction.
+   */
+  shouldFetchRootKey?: boolean;
 }
 
 function getDefaultFetch(): typeof fetch {
@@ -245,7 +250,9 @@ other computations so that this class can stay as simple as possible while
 allowing extensions.
  */
 export class HttpAgent implements Agent {
-  public rootKey: ArrayBuffer;
+  public rootKey: ArrayBuffer | null = null;
+  #rootKeyPromise: Promise<ArrayBuffer> | null = null;
+  #shouldFetchRootKey: boolean = false;
   #identity: Promise<Identity> | null;
   readonly #fetch: typeof fetch;
   readonly #fetchOptions?: Record<string, unknown>;
@@ -253,7 +260,6 @@ export class HttpAgent implements Agent {
   #timeDiffMsecs = 0;
   readonly host: URL;
   readonly #credentials: string | undefined;
-  #rootKeyFetched = false;
   readonly #retryTimes; // Retry requests N times before erroring by default
   #backoffStrategy: BackoffStrategyFactory;
   readonly #maxIngressExpiryInMinutes: number;
@@ -288,7 +294,8 @@ export class HttpAgent implements Agent {
     this.#fetch = options.fetch || getDefaultFetch() || fetch.bind(global);
     this.#fetchOptions = options.fetchOptions;
     this.#callOptions = options.callOptions;
-    this.rootKey = options.rootKey ? options.rootKey : fromHex(IC_ROOT_KEY);
+    this.#shouldFetchRootKey = options.shouldFetchRootKey ?? false;
+    this.rootKey = options.rootKey ?? null;
 
     const host = determineHost(options.host);
     this.host = new URL(host);
@@ -355,7 +362,7 @@ export class HttpAgent implements Agent {
   }
 
   public static async create(
-    options: HttpAgentOptions & { shouldFetchRootKey?: boolean } = {
+    options: HttpAgentOptions = {
       shouldFetchRootKey: false,
     },
   ): Promise<HttpAgent> {
@@ -435,6 +442,7 @@ export class HttpAgent implements Agent {
     },
     identity?: Identity | Promise<Identity>,
   ): Promise<SubmitResponse> {
+    await this.#rootKeyGuard();
     // TODO - restore this value
     const callSync = options.callSync ?? true;
     const id = await (identity !== undefined ? await identity : await this.#identity);
@@ -771,6 +779,7 @@ export class HttpAgent implements Agent {
     fields: QueryFields,
     identity?: Identity | Promise<Identity>,
   ): Promise<ApiQueryResponse> {
+    await this.#rootKeyGuard();
     const backoff = this.#backoffStrategy();
     const ecid = fields.effectiveCanisterId
       ? Principal.from(fields.effectiveCanisterId)
@@ -958,6 +967,7 @@ export class HttpAgent implements Agent {
     identity?: Identity | Promise<Identity>,
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
   ): Promise<any> {
+    await this.#rootKeyGuard();
     const id = await (identity !== undefined ? await identity : await this.#identity);
     if (!id) {
       throw new IdentityInvalidError(
@@ -995,6 +1005,7 @@ export class HttpAgent implements Agent {
     // eslint-disable-next-line
     request?: any,
   ): Promise<ReadStateResponse> {
+    await this.#rootKeyGuard();
     const canister = typeof canisterId === 'string' ? Principal.fromText(canisterId) : canisterId;
 
     const transformedRequest = request ?? (await this.createReadStateRequest(fields, identity));
@@ -1083,6 +1094,7 @@ export class HttpAgent implements Agent {
    * @param {Principal} canisterId - Pass a canister ID if you need to sync the time with a particular replica. Uses the management canister by default
    */
   public async syncTime(canisterId?: Principal): Promise<void> {
+    await this.#rootKeyGuard();
     const CanisterStatus = await import('../../canisterStatus');
     const callTime = Date.now();
     try {
@@ -1126,13 +1138,32 @@ export class HttpAgent implements Agent {
   }
 
   public async fetchRootKey(): Promise<ArrayBuffer> {
-    if (!this.#rootKeyFetched) {
-      const status = await this.status();
-      // Hex-encoded version of the replica root key
-      this.rootKey = (status as JsonObject & { root_key: ArrayBuffer }).root_key;
-      this.#rootKeyFetched = true;
+    // Wait for already pending promise
+    if (this.#rootKeyPromise) {
+      const result = await this.#rootKeyPromise;
+      this.#rootKeyPromise = null;
+      this.rootKey = result;
+      return result;
     }
-    return this.rootKey;
+    const status = await this.status();
+    // Hex-encoded version of the replica root key
+    const rootKey = (status as JsonObject & { root_key: ArrayBuffer }).root_key;
+    this.rootKey = rootKey;
+    rootKey;
+
+    return rootKey;
+  }
+
+  async #rootKeyGuard(): Promise<void> {
+    if (this.rootKey) {
+      return;
+    } else if (this.rootKey === null && this.#shouldFetchRootKey) {
+      await this.fetchRootKey();
+    } else {
+      throw new AgentError(
+        `Invalid root key detected. The root key for this agent is ${this.rootKey} and the shouldFetchRootKey value is set to ${this.#shouldFetchRootKey}. The root key should only be unknown if you are in local development. Otherwise you should avoid fetching and use the default IC Root Key or the known root key of your environment.`,
+      );
+    }
   }
 
   public invalidateIdentity(): void {
@@ -1144,6 +1175,7 @@ export class HttpAgent implements Agent {
   }
 
   public async fetchSubnetKeys(canisterId: Principal | string) {
+    await this.#rootKeyGuard();
     const effectiveCanisterId: Principal = Principal.from(canisterId);
     const response = await request({
       canisterId: effectiveCanisterId,